Archival, Lambda & Download updates

+++ app/functions/globals.py
@@ -2624,6 +2624,95 @@ def checkFileExist(file,folder):
         # file does not exist
         return None, False
 
+def checkFilesExist(files,folder):
+    '''Checks if a list of files exists in db & s3. Returns the filename if it does and None otherwise.'''
+
+    starttime = time.time()
+    already_uploaded = []  
+    req_lambda = []
+    new_names = {}
+
+    hash_dict = {}
+    file_dict = {}
+    camera_paths = []
+    
+    for file in files:
+        file_dict[file['hash']] = folder + '/' + file['name']
+        hash_dict[folder + '/' +file['name']] = file['hash']
+        camera_paths.append(folder + '/' + file['name'].rsplit('/',1)[0])
+
+
+    camera_paths = list(set(camera_paths))
+    common_path = folder + '/' + files[0]['name'].split('/')[0] + '/%'
+
+    hash_check = db.session.query(Image.hash).join(Camera).filter(Image.hash.in_(file_dict.keys())).filter(Camera.path.like(common_path))
+    vid_hash_check = db.session.query(Video.hash).join(Camera).filter(Video.hash.in_(file_dict.keys())).filter(Camera.path.like(common_path))
+    hash_check = hash_check.union(vid_hash_check).distinct().all()
+
+    for hash in hash_check:
+        filename = file_dict.get(hash[0])
+        if filename: already_uploaded.append(filename)
+
+    not_imported = list(set(file_dict.values()) - set(already_uploaded))
+
+    if not_imported:
+        in_s3 = []
+        in_s3_videos = []
+        other_s3 = []
+        isVideo = re.compile('(\.avi$)|(\.mp4$)|(\.mov$)', re.I)
+        for camera_path in camera_paths:
+            folders, filenames = list_all(Config.BUCKET,camera_path+'/',include_all=True)
+            for filename in filenames:
+                file = camera_path + '/' + filename
+                if file in not_imported:
+                    if isVideo.search(file):
+                        in_s3_videos.append(camera_path + '/_video_images_/' + filename.split('.')[0])
+                    in_s3.append(file)
+                elif '_(' in file:
+                    other_s3.append(file)
+
+        if in_s3:
+            filename_in_db = {d[1]+'/'+d[0]:d[2] for d in db.session.query(Image.filename,Camera.path, Image.hash)\
+                                        .join(Camera)\
+                                        .filter(Camera.path.like(common_path))\
+                                        .filter(Image.hash.notin_(file_dict.keys()))\
+                                        .filter((Camera.path + '/' + Image.filename).in_(in_s3))\
+                                        .distinct().all()}
+
+            if in_s3_videos:
+                vid_filename_in_db = db.session.query(Video.filename,Camera.path, Video.hash)\
+                                            .join(Camera)\
+                                            .filter(Camera.path.in_(in_s3_videos))\
+                                            .filter(Video.hash.notin_(file_dict.keys()))\
+                                            .distinct().all()
+                for d in vid_filename_in_db:
+                    filename_in_db[d[1].split('/_video_images_/')[0] + '/' + d[0]] = d[2]
+
+            for name, hash in filename_in_db.items():
+                fn_hash = hash_dict[name]
+                if fn_hash and fn_hash != hash:
+                    counter = 1 
+                    filename, ext = name.rsplit('.',1)
+                    new_name = filename + '_(' + str(counter) + ').' + ext
+                    while new_name in other_s3:
+                        counter += 1
+                        new_name =  filename + '_(' + str(counter) + ').' + ext
+                    new_names[name] = new_name.rsplit('/',1)[1]
+
+            req_lambda = list(set(in_s3) - set(new_names.keys()))
+            already_uploaded = list(set(already_uploaded+req_lambda))  
+
+    # Remove the folder name by splitting the first part of the path
+    already_uploaded = [x.split('/', 1)[1] for x in already_uploaded]
+    req_lambda = [x.split('/', 1)[1] for x in req_lambda]
+    new_names = {k.split('/', 1)[1]: v for k, v in new_names.items()}
+
+    if Config.DEBUGGING: 
+        app.logger.info('Checked files in {} seconds'.format(time.time()-starttime))
+        app.logger.info('Already uploaded: {}, Required Lambda: {}, Req new names: {}'.format(len(already_uploaded),len(req_lambda),len(new_names)))
+
+    return already_uploaded, req_lambda, new_names
+
 def rDets(sq):
     '''Adds the necessary SQLAlchemy filters for a detection to be considered 'relevent'. ie. non-static, not deleted and of sufficient confidence.'''
     return sq.filter(or_(and_(Detection.source==model,Detection.score>Config.DETECTOR_THRESHOLDS[model]) for model in Config.DETECTOR_THRESHOLDS))\
@@ -3567,21 +3656,24 @@ def manageDownload(task_id):
     if download_requests:
         cleanup = False
         for request in download_requests:
-            user_id = request.user_id
-            try:
-                download_params = json.loads(GLOBALS.redisClient.get('fileDownloadParams_'+str(task_id)+'_'+str(user_id)).decode())
-                survey_restore = request.task.survey.download_restore
-                expiry_date = calculate_restore_expiry_date(survey_restore,Config.RESTORE_TIME,Config.DOWNLOAD_RESTORE_DAYS)
-                if expiry_date and datetime.utcnow() < expiry_date:
-                    request.status = 'Available'
-                    request.timestamp = datetime.utcnow()
-                else:
-                    cleanup = download_params['include_empties']
+            if request.name == 'restore':
+                user_id = request.user_id
+                try:
+                    download_params = json.loads(GLOBALS.redisClient.get('fileDownloadParams_'+str(task_id)+'_'+str(user_id)).decode())
+                    survey_restore = request.task.survey.download_restore
+                    expiry_date = calculate_restore_expiry_date(survey_restore,Config.RESTORE_TIME,Config.DOWNLOAD_RESTORE_DAYS)
+                    if expiry_date and datetime.utcnow() < expiry_date:
+                        request.status = 'Available'
+                        request.timestamp = datetime.utcnow()
+                    else:
+                        cleanup = download_params['include_empties']
 
+                        GLOBALS.redisClient.delete('fileDownloadParams_'+str(task_id)+'_'+str(user_id))
+                        db.session.delete(request)
+                except:
                     GLOBALS.redisClient.delete('fileDownloadParams_'+str(task_id)+'_'+str(user_id))
                     db.session.delete(request)
-            except:
-                GLOBALS.redisClient.delete('fileDownloadParams_'+str(task_id)+'_'+str(user_id))
+            else:
                 db.session.delete(request)
 
         if cleanup:
